refactor(auth): fold toApiAccessScope into enforceAuth and relocate to public-api/server
